Don't force the fd to be active in net_wakeup() net_wakeup() forces the file descriptor it's waking up to be active, by setting active_fd[fd] = 1.
This can cause the following race:

 * Lock L is held by client A
 * Client B is connected and waiting for lock L to become available
 * Client B decides to disconnect, and closes the socket
 * This triggers net_close()
 * At this point, client A releases lock L, triggering net_wakeup()

Now, here's how it could go:

  net_close()              net_wakeup()
    active_fd[fd] = 0
                             active_fd[fd] = 1
                             net_send_cmd[fd] returns OK because
                               the socket is not closed yet
    close()


And that leaves fd closed, but active_fd[fd] = 1, which will cause the
select() loop to hang up indefinitely, because it gets a "Bad file descriptor"
error (due to a fd on the read set being closed), which causes it to rebuild
the read set and try again. But nothing will never remove that fd from the
set, because nothing will set active_fd[fd] = 0.

The fix is not setting active_fd[fd] = 1 in net_wakeup(), it has nothing
to do there in the first place.


Besides, check if the fd is active before calling net_send_cmd(), because we
know it's going to fail.

Thanks to Lloyd C. Cha for the report and testing.